BE-#13 refactor : Review 수정
- Review 내용을 바탕으로 코드 수정

-------------------------
-------------------------
