+++ BE/src/main/java/kr/codesquad/issuetracker/common/config/WebConfig.java
@@ -1,5 +1,8 @@
 package kr.codesquad.issuetracker.common.config;
 
+import kr.codesquad.issuetracker.common.security.GithubKey;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
@@ -11,6 +14,17 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
+
+    @Value("${GITHUB_CLIENT_ID}")
+    private String GITHUB_CLIENT_ID;
+
+    @Value("${GITHUB_CLIENT_SECRET}")
+    private String GITHUB_CLIENT_SECRET;
+
+    @Bean
+    public GithubKey githubKey() {
+        return new GithubKey(GITHUB_CLIENT_ID,GITHUB_CLIENT_SECRET);
+    }
 
     @Override
     public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
